requirements.txt
stanza
flask
flask_restful
python-dotenv
pandas
sparqlwrapper
gender_guesser
pillow